Fix flaky testConsumerEventWithoutPartition cause by the change of Pulsar 3.0
### Motivation

https://github.com/apache/pulsar/pull/19502 brings a change to the
consumer selection of the Failover subscription.

Before that PR, if a new consumer joins a Failover subscription when
there is a consumer that subscribes the topic, the new consumer will
receive the "inactive" event.

After that PR, a random consumer will receive the "inactive" event.

### Modifications

Change the assertion when `consumer` subscribes the topic with the same
subscription.

In addition, use a thread safe queue to avoid race conditions.