Close the socket gracefully on Windows (#284)
* Close the socket gracefully on Windows

Fixes https://github.com/apache/pulsar-client-cpp/issues/261

### Motivation

When closing the socket on Windows, the `ERROR_CONNECTION_ABORTED` error
might be returned in the callback of `async_receive`. It's caused by the
`socket::close` method is not portable enough, see
https://www.boost.org/doc/libs/1_81_0/doc/html/boost_asio/reference/basic_stream_socket/close/overload2.html

> For portable behaviour with respect to graceful closure of a connected socket, call shutdown() before closing the socket.

### Modifications

Call `shutdown` in `ClientConnection::closeSocket` before calling
`close`. In addition, when no bytes are returned or the `eof` error
happened,  print the logs with debug level whatever the error code
is in the read callback.

* Handle operation_failed first