Fix build failure with the Protobuf 23.3 (#290)
### Motivation

The current CMakeLists.txt does not work with the latest Protobuf
(23.3.0). It's because currently the Module mode is used to find
Protobuf, while the `FindProtobuf.cmake` is not updated to find the
Abseil dependency.

See https://github.com/protocolbuffers/protobuf/issues/12292#issuecomment-1529680040

### Modifications

For macOS, use the Config mode to find Protobuf. It's because in other
systems, the Module mode works well. Besides, enable `PROTOBUF_USE_DLLS`
as a workaround for
https://github.com/protocolbuffers/protobuf/pull/12983 is not released.

Pin the default C++ standard to 17 for macOS so that users don't need to
set the C++ standard manually.