Fix the buggy Future and Promise implementations (#299)
Fixes https://github.com/apache/pulsar-client-cpp/issues/298

### Motivation

Currently the `Future` and `Promise` are implemented manually by
managing conditional variables. However, the conditional variable
sometimes behaviors incorrectly on macOS, while the existing `future`
and `promise` from the C++ standard library works well.

### Modifications

Redesign `Future` and `Promise` based on the utilities in the standard
`<future>` header. In addition, fix the possible race condition when
`addListener` is called after `setValue` or `setFailed`:
- Thread 1: call `setValue`, switch existing listeners and call them one
  by one out of the lock.
- Thread 2: call `addListener`, detect `complete_` is true and call the
  listener directly.

Now, the previous listeners and the new listener are called concurrently
in thread 1 and 2.

This patch fixes the problem by adding a future to wait all listeners
that were added before completing are done.

### Verifications

Run the reproduce code in #298 for 10 times and found it never failed or
hang.

Co-authored-by: Zike Yang <zike@apache.org>

---------

Co-authored-by: Zike Yang <zike@apache.org>