[fix] callback of send batch message is error when flush (#303)
### Motivation

When sendAsync batch messages and then flush() and close(), found that the callback of message is ResultAlreadyClosed but not Ok. 

The root is if there is no batch messages remain in BatchMessageContainer when do flush,  the flush callback directly return Ok. Instead, it should return Ok until the lastSendOpFuture complete.  

### Modifications

1. fix the error code.
2. test two cases of flush batch messages.