Fix the handler instance is expired when the connection is established
### Motivation

We observed some logs that showed the handler instance is expired when
the connection is established after a reconnection:

```
HandlerBase Weak reference is not valid anymore
```

https://github.com/apache/pulsar-client-cpp/blob/b35ae1aa4b9834886c0889635de81834f9b2f774/lib/HandlerBase.cc#L92

### Modifications

Pass a `shared_ptr` instead of a `weak_ptr` in `HandlerBase::grabCnx` to
ensure the `connectionOpened` or `connectionFailed` callback is called
if `ClientImpl::getConnection` is called. We only need to pass a
`weak_ptr` in `scheduleReconnection` to skip the reconnection if the
handler is expired.