Add the Consumer::getConsumerName API
### Motivation

C++ client supports setting the consumer name via
`ConsumerConfiguration::setConsumerName` but it does not support getting
the consumer name for a given consumer.

### Modifications

Add the `getConsumerName` method to `ConsumerImplBase` to get the
consumer name from the config and a public method with the same name to
`Consumer`.

Add `ConsumerTest.testConsumerName` to verify the consumer name is set
correctly in the subscribe command.