Fix the operation timeout is not honored for GetSchema requests
Fixes https://github.com/apache/pulsar-client-cpp/issues/357

### Motivation

In `Client::getSchemaInfoAsync`, the underlying
`LookupService::getSchema` is called, which does not register a timer
according to the operation timeout.

### Modifications

Register a timer and cache it and the GetSchema promise in
`ClientConnection::pendingGetSchemaRequests_`, then remove the entry and
complete the promise with `ResultTimeout`.

To verify the timeout is honored, modify the timeout type from `int` to
`chrono::nanoseconds` so that a smaller timeout can be configured. Then
test a small enough timeout (1 ns) in
`LookupServiceTest.testGetSchemaTimeout`.