Fix segmenatation fault if connection is closed during handshake
Fixes https://github.com/apache/pulsar-client-cpp/issues/398

### Motivation

There is an implicit requirement for the `async_handshake` API that
before it's done, the `socket` and `ssl::stream` objects must exist.
Otherwise segmenatation fault might happen.

### Modifications

Capture `socket_` and `tlsSocket_` in the handshake callback. It can be
verified by modifying the code according to #398 and rerun the test.