Support customize vcpkg directory when INTEGRATE_VCPKG is ON (#417)
### Motivation

Currently when INTEGRATE_VCPKG is ON, the CMAKE_TOOLCHAIN_FILE variable
is always a subdirectory of `${CMAKE_SOURCE_DIR}/vcpkg`. We can only
customize the vcpkg directory when INTEGRATE_VCPKG is OFF, while the
legacy CMake logic is incompatible with this way.

### Modifications

When INTEGRATE_VCPKG is ON, only set CMAKE_TOOLCHAIN_FILE if it's not
defined. The workflow and README are updated for it.

(cherry picked from commit 821871777e247e1ccbfa323ea0d5136cf0e18711)