Handle the exception from the token supplier
### Motivation

When a token supplier is passed to the `AuthToken`, if exceptions are
thrown from it, the application will crash immediately. A typical case
is the Python wrapper might raise an exception when trying to get token.

### Modifications

Catch the exception in `Commands::newConnect` because the token supplier
is called in it. Then convert it to `ResultAuthenticationError`. Add
`testTokenSupplierException` to verify it.